requirements.txt
Django==3.2.16
djangorestframework==3.11.2
drf-jwt==1.19.2